requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
streamlit>=1.40.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
matplotlib>=3.8.0
chromadb>=0.5.0
pypdf>=4.0.0
python-docx>=1.1.0
